# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/cabinets": {
            "get": {
                "tags": [
                    "Cabinet"
                ],
                "summary": "CabinetsInfo",
                "operationId": "e857b0cb4a07f10bbbd0f1b6706286b7",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cabinets": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-12-20 17:49:36"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-12-20 17:49:36"
                                                    },
                                                    "purpose": {
                                                        "type": "string",
                                                        "example": "Russian"
                                                    },
                                                    "number": {
                                                        "type": "string",
                                                        "example": "21"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cabinet"
                ],
                "summary": "Create",
                "operationId": "a30829548ca3d941c88abb33c242cc6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number",
                                    "purpose"
                                ],
                                "properties": {
                                    "number": {
                                        "type": "string",
                                        "example": "21"
                                    },
                                    "purpose": {
                                        "type": "string",
                                        "example": "Russian"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cabinet": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "purpose": {
                                                    "type": "string",
                                                    "example": "Russian"
                                                },
                                                "number": {
                                                    "type": "string",
                                                    "example": "21"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/cabinets/{cabinet}": {
            "get": {
                "tags": [
                    "Cabinet"
                ],
                "summary": "GetCabinetID",
                "operationId": "4b4621ebbb2a0c03a23f914a70ac2c60",
                "parameters": [
                    {
                        "name": "cabinet",
                        "in": "path",
                        "description": "Cabinet ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cabinet": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "purpose": {
                                                    "type": "string",
                                                    "example": "Russian"
                                                },
                                                "number": {
                                                    "type": "string",
                                                    "example": "21"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cabinet"
                ],
                "summary": "Update",
                "operationId": "82ed994e47f5944e478c3a8a40fbda08",
                "parameters": [
                    {
                        "name": "cabinet",
                        "in": "path",
                        "description": "Cabinet ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number",
                                    "purpose"
                                ],
                                "properties": {
                                    "number": {
                                        "type": "string",
                                        "example": "21"
                                    },
                                    "purpose": {
                                        "type": "string",
                                        "example": "Russian"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cabinet": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "purpose": {
                                                    "type": "string",
                                                    "example": "Russian"
                                                },
                                                "number": {
                                                    "type": "string",
                                                    "example": "21"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cabinet"
                ],
                "summary": "Delete",
                "operationId": "a00a41c9bb166c77b990294ed76a048d",
                "parameters": [
                    {
                        "name": "cabinet",
                        "in": "path",
                        "description": "Cabinet ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}